refactor(GN-55): 알림 api 요청 함수 apis 디렉토리로 이동
